+++ packages/iov-bns-governance/src/governor.spec.ts
@@ -1,7 +1,9 @@
+/* eslint-disable @typescript-eslint/camelcase */
 import { ReadonlyDate } from "readonly-date";
 
-import { Address, Identity, TokenTicker } from "@iov/bcp";
+import { Address, Algorithm, Identity, PubkeyBytes, TokenTicker } from "@iov/bcp";
 import { ActionKind, bnsCodec, BnsConnection, VoteOption } from "@iov/bns";
+import { Encoding } from "@iov/encoding";
 import { Ed25519HdWallet, HdPaths, UserProfile } from "@iov/keycontrol";
 
 import { CommitteeId } from "./committees";
@@ -222,6 +224,91 @@ describe("Governor", () => {
           },
         },
         description: "Proposal to remove abcd from committee 5",
+        electionRuleId: 1,
+        startTime: 1562164525,
+        author: bnsCodec.identityToAddress(options.identity),
+        fee: {
+          tokens: {
+            quantity: "10000000",
+            fractionalDigits: 9,
+            tokenTicker: "CASH" as TokenTicker,
+          },
+        },
+      });
+
+      options.connection.disconnect();
+    });
+
+    it("works for AddValidator", async () => {
+      pendingWithoutBnsd();
+      const options = await getConnectionAndIdentity();
+      const governor = new Governor(options);
+
+      const tx = await governor.buildCreateProposalTx({
+        type: ProposalType.AddValidator,
+        title: "Add abcd as validator",
+        description: "Proposal to add abcd as validator",
+        startTime: new ReadonlyDate(1562164525898),
+        electionRuleId: 1,
+        pubkey: {
+          algo: Algorithm.Ed25519,
+          data: Encoding.fromHex("abcd") as PubkeyBytes,
+        },
+        power: 12,
+      });
+      expect(tx).toEqual({
+        kind: "bns/create_proposal",
+        creator: options.identity,
+        title: "Add abcd as validator",
+        action: {
+          kind: ActionKind.SetValidators,
+          validatorUpdates: {
+            ed25519_abcd: { power: 12 },
+          },
+        },
+        description: "Proposal to add abcd as validator",
+        electionRuleId: 1,
+        startTime: 1562164525,
+        author: bnsCodec.identityToAddress(options.identity),
+        fee: {
+          tokens: {
+            quantity: "10000000",
+            fractionalDigits: 9,
+            tokenTicker: "CASH" as TokenTicker,
+          },
+        },
+      });
+
+      options.connection.disconnect();
+    });
+
+    it("works for RemoveValidator", async () => {
+      pendingWithoutBnsd();
+      const options = await getConnectionAndIdentity();
+      const governor = new Governor(options);
+
+      const tx = await governor.buildCreateProposalTx({
+        type: ProposalType.RemoveValidator,
+        title: "Remove validator abcd",
+        description: "Proposal to remove validator abcd",
+        startTime: new ReadonlyDate(1562164525898),
+        electionRuleId: 1,
+        pubkey: {
+          algo: Algorithm.Ed25519,
+          data: Encoding.fromHex("abcd") as PubkeyBytes,
+        },
+      });
+      expect(tx).toEqual({
+        kind: "bns/create_proposal",
+        creator: options.identity,
+        title: "Remove validator abcd",
+        action: {
+          kind: ActionKind.SetValidators,
+          validatorUpdates: {
+            ed25519_abcd: { power: 0 },
+          },
+        },
+        description: "Proposal to remove validator abcd",
         electionRuleId: 1,
         startTime: 1562164525,
         author: bnsCodec.identityToAddress(options.identity),