+++ packages/iov-bns/src/bnsconnection.spec.ts
@@ -1901,6 +1901,21 @@ describe("BnsConnection", () => {
     });
   });
 
+  describe("getValidators", () => {
+    it("works", async () => {
+      pendingWithoutBnsd();
+      const connection = await BnsConnection.establish(bnsdTendermintUrl);
+
+      const validators = await connection.getValidators();
+      expect(validators.length).toEqual(1);
+      expect(validators[0].pubkey.algo).toEqual(Algorithm.Ed25519);
+      expect(validators[0].pubkey.data.length).toEqual(32);
+      expect(validators[0].power).toEqual(10);
+
+      connection.disconnect();
+    });
+  });
+
   describe("getElectorates", () => {
     it("can query electorates set in genesis", async () => {
       pendingWithoutBnsd();