+++ packages/iov-bns/src/encode.spec.ts
@@ -1,3 +1,5 @@
+import Long from "long";
+
 import {
   Address,
   Algorithm,
@@ -525,7 +527,7 @@ describe("Encode", () => {
     // Governance
 
     describe("CreateProposalTx", () => {
-      it("works for CreateProposalTx with CreateTextResolution action", () => {
+      it("works with CreateTextResolution action", () => {
         const createProposal: CreateProposalTx & WithCreator = {
           kind: "bns/create_proposal",
           creator: defaultCreator,
@@ -556,7 +558,7 @@ describe("Encode", () => {
         });
       });
 
-      it("works for CreateProposalTx with UpdateElectorate action", () => {
+      it("works with UpdateElectorate action", () => {
         const createProposal: CreateProposalTx & WithCreator = {
           kind: "bns/create_proposal",
           creator: defaultCreator,
@@ -582,6 +584,47 @@ describe("Encode", () => {
               metadata: { schema: 1 },
               electorateId: fromHex("0000000000000005"),
               diffElectors: [{ address: fromHex("6e1114f57410d8e7bcd910a568c9196efc1479e4"), weight: 8 }],
+            },
+          }).finish(),
+          description: "foo bar",
+          electionRuleId: fromHex("0011221122112200"),
+          startTime: 1122334455,
+          author: fromHex("6e1114f57410d8e7bcd910a568c9196efc1479e4"),
+        });
+      });
+
+      it("works with SetValidators action", () => {
+        const createProposal: CreateProposalTx & WithCreator = {
+          kind: "bns/create_proposal",
+          creator: defaultCreator,
+          title: "Why not try this?",
+          action: {
+            kind: ActionKind.SetValidators,
+            validatorUpdates: {
+              // eslint-disable-next-line @typescript-eslint/camelcase
+              ed25519_0902bb5de30ccb15b6decb6aa1fdb4f0c1c7317df62dcafa81ccad82ce88dd22: { power: 5 },
+            },
+          },
+          description: "foo bar",
+          electionRuleId: 4822531585417728,
+          startTime: 1122334455,
+          author: defaultSender,
+        };
+        const msg = buildMsg(createProposal).govCreateProposalMsg!;
+        expect(msg).toEqual({
+          metadata: { schema: 1 },
+          title: "Why not try this?",
+          rawOption: codecImpl.bnsd.ProposalOptions.encode({
+            validatorsApplyDiffMsg: {
+              metadata: { schema: 1 },
+              validatorUpdates: [
+                {
+                  pubKey: {
+                    data: fromHex("0902bb5de30ccb15b6decb6aa1fdb4f0c1c7317df62dcafa81ccad82ce88dd22"),
+                  },
+                  power: Long.fromNumber(5),
+                },
+              ],
             },
           }).finish(),
           description: "foo bar",